extract telemetry handler module from unit tests
